Add files via upload
This project is a e-commerce clone of the Amazon front page, by using only HTML and CSS. It aims to replicate the layout, styling, and basic structure of Amazon's homepage to provide a visually similar experience like the real one. This project is a part of my learning journey to master web development, focusing on HTML and CSS fundamentals without involving any JavaScript or backend functionalities. This is my first ever cloning project.